fix(busca): busca global retorna resultados; QuantidadeAtual decimal (QA v1.10 BUG-01)
A busca do header (Ctrl+K) voltava vazia para todo termo com saldo. O DTO Web
ResultadoBuscaUnificada.QuantidadeAtual era int?, mas a API serializa o saldo de
itens_estoque/movimentacoes_estoque (numeric(18,3)) com escala (ex.: 25.000), o
que lancava JsonException ao desserializar o List<> inteiro -> ApiClient devolvia
PARSE_ERROR -> BuscaController mascarava como []. Agora QuantidadeAtual e decimal?
(bate com o contrato ResultadoBuscaInteligente) e o BuscaController separa falha
real (logada) de "nenhum resultado". Teste de regressao em Web.UnitTests.

refs #664

+++ EasyStock.Web/Models/Api/ResultadoBuscaUnificada.cs
@@ -11,7 +11,10 @@ public record ResultadoBuscaUnificada
     public required string ChaveExibicao { get; init; }
     public int Score { get; init; }
     public string? Sku { get; init; }
-    public int? QuantidadeAtual { get; init; }
+    // numeric(18,3) na API (itens_estoque.QuantidadeAtual / movimentacoes_estoque.Quantidade):
+    // o JSON vem com escala (ex.: 25.000). Tinha que ser decimal? — int? quebrava a
+    // desserializacao do List<> inteiro -> PARSE_ERROR -> busca global vazia (QA v1.10 BUG-01).
+    public decimal? QuantidadeAtual { get; init; }
     public string? Status { get; init; }
     public string? FornecedorNome { get; init; }
     public string? Loja { get; init; }
